gio_v3/requirements.txt
flask==3.1.3
werkzeug==3.1.6
jinja2==3.1.6
click==8.3.1
markupsafe==3.0.3
itsdangerous==2.2.0
gunicorn==23.0.0
python-dotenv==1.0.1
tzdata
pytest>=8.0
pytest-flask>=1.3
